// frontend-admin/src/api/generated/model/backupArtifactResponseDtoArtifactType.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * Kasse API
 * Registrierkasse API - RKSV uyumlu kasa sistemi
 * OpenAPI spec version: v1
 */

export type BackupArtifactResponseDtoArtifactType = typeof BackupArtifactResponseDtoArtifactType[keyof typeof BackupArtifactResponseDtoArtifactType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const BackupArtifactResponseDtoArtifactType = {
  NUMBER_0: 0,
  NUMBER_1: 1,
  NUMBER_2: 2,
  NUMBER_3: 3,
  NUMBER_4: 4,
  NUMBER_5: 5,
} as const;